Test that complex secrets are provided to k8s
Instead of testing that the secret `supersecret` is provided successfully to k8s, we now test more complex secrets, such as JSON Objects and SSH Keys. We don't test also simple secrets as if complex one are provided then simple one will be as well.

We run the tests in different files for 2 reasons:
  1. In case of failure it is clear which complex secret failed.
  2. We need to tear down and re-create the environment in order to test
     these cases so the secrets-provider pod is re-created.